rename: AnnouncementImgContainer -> AnnounceImg로 파일명 변경
